agents/requirements.txt
flask==2.3.2
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
gunicorn==20.1.0
flask-cors==3.0.10
psycopg2-binary==2.9.7
joblib==1.3.2
